auth-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
click>=8.1.7
psycopg2-binary>=2.9.0
